refactor: Hoist editor provider to editor
Prepare for text editor.
